requirements.txt
anytree==2.2.1
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
colorama==0.3.7
data-importer==3.0.2
DateTime==4.2
decorator==4.0.11
dj-database-url==0.4.2
django==1.10
django-debug-toolbar==1.8
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
et-xmlfile==1.0.1
Events==0.3
ez-setup==0.9
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.5
imageio==2.1.2
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
locustio==0.8a2
MarkupSafe==1.0
moviepy==0.2.3.2
msgpack-python==0.4.8
music21==3.1.0
MySQL-python==1.2.5
mysqlclient==1.3.10
numpy==1.13.0
oauthlib==2.0.2
olefile==0.44
openpyxl==2.4.0
pgxnclient==1.2.1
Pillow==4.2.0
profilehooks==1.9.0
psutil==5.2.2
psycopg2==2.7.1
py==1.4.31
pygame==1.9.2a0
Pygments==2.2.0
PyJWT==1.5.0
Pympler==0.5
PyMySQL==0.7.11
pypiwin32==219
pytest==2.9.1
pytz==2017.2
pyzmq==15.2.0
requests==2.18.1
six==1.10.0
sqlparse==0.2.3
tqdm==4.11.2
unroll==0.1.0
urllib3==1.21.1
virtualenv==15.1.0
Werkzeug==0.12.2
whitenoise==3.3.0
WMI==1.4.9
xlrd==1.0.0
zope.interface==4.4.2
